+++ src/Utilities.cc
@@ -6,12 +6,12 @@ namespace Hexit
 {
 namespace fs = std::filesystem;
 
-bool validate_args(std::size_t argc, const char* const* const argv)
+bool validate_args(std::uintmax_t argc, const char* const* const argv)
 {
-    std::size_t i      = 0u;
-    bool        help   = false;
-    bool        file   = false;
-    bool        offset = false;
+    std::uintmax_t i      = 0u;
+    bool           help   = false;
+    bool           file   = false;
+    bool           offset = false;
     for (; i < argc && argv[i];)
     {
         std::string sarg(argv[i]);
@@ -70,7 +70,7 @@ bool get_flag(int argc, const char* const* const argv, const std::string& flag)
     return res != (argv + argc);
 }
 
-std::uint64_t str_to_int(const char* const str)
+std::uintmax_t str_to_int(const char* const str)
 {
     if (!str)
         return 0;
